refactor: Move WooPay eligibility endpoint to dedicated WooPayments REST controller
